removed old alias code
