+++ hoprd/db/api/Cargo.toml
@@ -20,12 +20,9 @@ chrono = { workspace = true }
 bincode = "1.3.3"
 futures = { workspace = true }
 lazy_static = { workspace = true, optional = true }
-libp2p-identity = { workspace = true }
-moka = { workspace = true }
-multiaddr = { workspace = true }
-num_enum = { workspace = true }
 sea-orm = { workspace = true }
 sea-query = { version = "0", default-features = false, features = ["with-chrono", "derive", "backend-sqlite", "attr"]}
+serde = { workspace = true, features = ["derive"] }
 serde_json = { workspace = true }
 sea-query-binder = { version = "0",  default-features = false, features = ["with-chrono", "sqlx-sqlite", "runtime-async-std-rustls"]}
 smart-default = { workspace = true }
@@ -34,9 +31,8 @@ strum = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
 
-hopr-crypto-types = { workspace = true}
-hopr-crypto-packet = { workspace = true}
-hopr-primitive-types = { workspace = true}
+hopr-crypto-types = { workspace = true }
+hopr-primitive-types = { workspace = true }
 hoprd-db-entity = { workspace = true }
 hoprd-db-migration = { workspace = true }
 hopr-internal-types = { workspace = true }
@@ -45,7 +41,6 @@ hopr-metrics = { workspace = true, optional = true }
 [dev-dependencies]
 async-std = { workspace = true, features = ["attributes"] }
 env_logger = { workspace = true }
-rand = "0.8.5"
 lazy_static = { workspace = true }
 hopr-crypto-random = { workspace = true }
 hex-literal = { workspace = true }
