+++ hoprd/db/entity/Cargo.toml
@@ -24,6 +24,6 @@ sea-orm = { workspace = true }
 chrono = { workspace = true }
 thiserror = { workspace = true }
 
-hopr-crypto-types = { workspace = true }
+hopr-crypto-types = { workspace = true}
 hopr-internal-types = { workspace = true }
 hopr-primitive-types = { workspace = true }
